switched freematrix(double) to integer arguments

+++ src/grid.h
@@ -30,8 +30,8 @@ double **matrixreservdouble(int nrl, int nrh, int ncl, int nch);
 /*******************************************************/
 /* SET MARTIX-ARRAY TO ZERO                            */
 /*******************************************************/
-void freematrix(float **m,long nrl, long nrh, long ncl, long nch);
-void freematrixdouble(double **m,long nrl, long nrh, long ncl, long nch);
+void freematrix(float **m,int nrl, int nrh, int ncl, int nch);
+void freematrixdouble(double **m,int nrl, int nrh, int ncl, int nch);
 
 void initializeglacier2zero_nodata(int nrows, int ncols, float **glaciergrid);
